requirements.txt
Django~=3.2.12
gunicorn==20.1.0
psycopg2-binary
django-rest-framework
fcm-django==1.0.5
firebase-admin==5.0.3
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.7.1
social-auth-app-django
djangorestframework~=3.13.1
django-simple-history==3.0.0
Pillow==9.1.0
gusregon==1.1.3
django-crispy-forms==1.14.0

requests~=2.27.1
